Move snapshotstore under store
