+++ store/snapshots/chunk_test.go
