+++ store/snapshots/helpers_test.go
