+++ store/snapshots/manager_test.go
