+++ store/snapshots/store_test.go
