+++ store/snapshots/stream.go
