+++ store/snapshots/types/errors.go
