+++ store/snapshots/types/format.go
