+++ store/snapshots/types/options.go
