+++ store/snapshots/types/util.go
